requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0
matplotlib>=3.8
SQLAlchemy>=2.0
aiosqlite>=0.19
pandas>=2.2
numpy>=1.26
ta>=0.11
tqdm>=4.66
streamlit>=1.34
plotly>=5.20